+++ packages/@codotype/ui/src/components/BlueprintEditor/components/NewModalButton.vue
@@ -4,7 +4,7 @@
     class="btn btn-primary btn-sm"
     @click="$store.dispatch(vuexAction, true)"
   >
-    <i class="fa fa-plus"></i>
+    <i class="fa fa-plus" />
   </button>
 </template>