+++ packages/@codotype/ui/src/components/BlueprintEditor/index.vue
@@ -1,37 +1,26 @@
 <template>
-  <div class="card card-body shadow-sm">
+  <div class="card card-body shadow-sm mb-4">
     <b-row class='d-flex align-items-center'>
       <b-col sm=6>
         <span class='d-flex align-items-center'>
           <h4 class="mb-0 mr-2 d-flex">{{projectName}}</h4>
-          <i class='fa fa-fw fa-pencil-alt text-muted' @click="$store.dispatch('editor/project/edit')" />
+          <ProjectEditButton />
           <EditProjectModal />
+          <HelpPopover
+            target="projectEditPopover"
+            placement="right"
+            :triggers="['hover']"
+            content='Edit Project Name'>
+          </HelpPopover>
         </span>
       </b-col>
       <b-col sm=6 class='d-flex justify-content-end'>
-        <b-dropdown
-          no-caret
-          size="sm"
-          variant="light"
-          class='mr-2'
-          toggle-class='rounded'
-        >
-          <template slot="button-content">
-            <i class="fa fa-fw fa-ellipsis-h" />
-          </template>
-
-          <b-dropdown-item-button @click="$store.commit('editor/modals/import/showing', true)">
-            <i class="fa fa-fw fa-upload"></i>
-            Import Project
-          </b-dropdown-item-button>
-
-          <b-dropdown-item-button @click="$store.commit('editor/modals/export/showing', true)">
-            <i class="fa fa-fw fa-download"></i>
-            Export Project
-          </b-dropdown-item-button>
-
-        </b-dropdown>
+        <HelpButton />
+        <!-- <TourButton /> -->
+        <ProjectDropdown />
         <GenerateCodeButton />
+        <ImportModal />
+        <ExportModal />
       </b-col>
       <b-col sm=12>
         <hr />
@@ -41,40 +30,20 @@
     <b-row>
       <b-col sm=12>
         <b-tabs>
-          <b-tab title="Schemas" lazy active>
-            <b-row class='mt-4 d-flex justify-content-center' v-if="!schemas[0]">
-              <b-col sm=10>
-                <SchemaEmptyState />
-              </b-col>
-            </b-row>
-            <b-row class='mt-3' v-else>
-              <b-col xl=3 lg=3 sm=12 class='border-right'>
-                <ImportModal />
-                <ExportModal />
-
-                <SchemaNewModal />
-                <SchemaNewButton />
-
-                <HelpPopover
-                  target="new-schema-button"
-                  placement="bottom"
-                  content='Create New Schema'>
-                </HelpPopover>
-
-                <SchemaList />
-              </b-col>
-
-              <b-col xl=9 lg=9 sm=12>
-                <SchemaDetail />
-              </b-col>
-            </b-row>
+          <b-tab
+            title="Schemas"
+            lazy
+            active
+            v-if="!model.self_configuring"
+          >
+            <SchemaEditor />
           </b-tab>
 
           <b-tab
             lazy
             class='pt-0'
+            v-for="group, index in model.configuration_groups"
             :title="group.label_plural || group.label"
-            v-for="group in model.configuration_groups"
             :key="group.identifier"
           >
             <b-row class="justify-content-center mt-3">
@@ -122,19 +91,19 @@ import { mapGetters, mapActions } from 'vuex'
 
 import ImportModal from './components/ImportModal'
 import ExportModal from './components/ExportModal'
-import SchemaNewButton from './components/schema/SchemaNewButton'
-import SchemaNewModal from './components/schema/SchemaNewModal'
-import SchemaList from './components/schema/SchemaList'
-import SchemaDetail from './components/schema/SchemaDetail'
-import SchemaEmptyState from './components/schema/SchemaEmptyState'
-import HelpPopover from '../HelpPopover'
+import SchemaEditor from './components/schema/SchemaEditor'
 import GenerateCodeButton from '../../modules/build/components/GenerateCodeButton'
+import HelpButton from '../HelpButton'
+import TourButton from '../TourButton'
+import HelpPopover from '../HelpPopover'
 import EditorHeader from '../EditorHeader'
 import GlobalOptionEditor from '../../modules/build/components/ConfigurationEditor/components/GlobalOptionEditor'
 import ModelOptionEditor from '../../modules/build/components/ConfigurationEditor/components/ModelOptionEditor'
 import ModelAddonEditor from '../../modules/build/components/ConfigurationEditor/components/ModelAddonEditor'
 import GlobalAddonEditor from '../../modules/build/components/ConfigurationEditor/components/GlobalAddonEditor'
 import EditProjectModal from './components/project/EditProjectModal'
+import ProjectEditButton from './components/project/ProjectEditButton'
+import ProjectDropdown from './components/project/ProjectDropdown'
 
 import {
   CONFIGURATION_GROUP_TYPE_OPTION,
@@ -151,21 +120,21 @@ export default {
     }
   },
   components: {
+    HelpPopover,
+    HelpButton,
+    TourButton,
     ImportModal,
     ExportModal,
-    SchemaNewButton,
-    SchemaNewModal,
-    HelpPopover,
-    SchemaList,
-    SchemaDetail,
-    SchemaEmptyState,
+    SchemaEditor,
     GenerateCodeButton,
     EditorHeader,
     GlobalOptionEditor,
     ModelOptionEditor,
     ModelAddonEditor,
     GlobalAddonEditor,
-    EditProjectModal
+    EditProjectModal,
+    ProjectEditButton,
+    ProjectDropdown
   },
   created () {
     this.$store.dispatch('editor/created')