+++ packages/@codotype/ui/src/components/HelpButton.vue
@@ -1,11 +1,10 @@
 <template>
   <span
-    class="text-muted"
+    class="badge text-muted bg-transparent mr-2"
     @mouseover="$store.commit('editor/help/showing', true)"
     @mouseout="$store.commit('editor/help/showing', false)"
   >
-    <i class="fa fa-lg fa-question-circle"></i>
-    Help
+    <i class="fa fa-fw fa-2x fa-question-circle" />
   </span>
 </template>