+++ packages/@codotype/ui/CHANGELOG.md
@@ -1,4 +1,4 @@
-### v0.1.13 (UNPUBLISHED)
+### v0.1.13
 
 #### Changes
 * Fixed another issue with `SchemaDestroyButton` tooltip behavior
@@ -7,6 +7,9 @@
 * Refactored `EditProjectModal` component, wraps `ProjectForm` component
 * Sized down `GenerateCodeButton`
 * Merged `ConfigurationEditor` component template and logic into `BlueprintEditor` component
+* Added `ProjectEditButton` component
+* Split `SchemaEditor` out of `BlueprintEditor` as a standalone component
+* Cleaned up `<TourButton />` component
 
 ---